+++ modules/app-export/src/main/java/com/axway/apim/appexport/lib/AppExportCLIOptions.java
@@ -13,41 +13,42 @@ public class AppExportCLIOptions extends StandardExportCLIOptions {
 	public AppExportCLIOptions(String[] args) throws ParseException {
 		super(args);
 		// Define command line options required for Application export
-		Option option = new  Option("n", "name", true, "Name of the application to export. You may use wildcards at the end or beginning. (e.g. *App*).");
+		Option option = new  Option("n", "name", true, "Filter applications with the specified name. You may use wildcards at the end or beginning.");
 		option.setRequired(false);
+		option.setArgName("*My Great App*");
 		options.addOption(option);
 		
-		option = new  Option("id", true, "The ID of the application.");
+		option = new  Option("id", true, "Filter the export to an application with that specific ID.");
 		option.setRequired(false);
+		option.setArgName("UUID-ID-OF-THE-APP");
 		options.addOption(option);
 
-		option = new  Option("state", true, "Export application with specific state: pending | approved");
+		option = new  Option("state", true, "Filter applications to the specificied state: pending | approved");
 		option.setRequired(false);
+		option.setArgName("pending");
 		options.addOption(option);
 
-		option = new  Option("orgName", true, "Limit applications to this organization");
+		option = new  Option("orgName", true, "Filter applications to this organization");
 		option.setRequired(false);
-		options.addOption(option);
-
-		option = new Option("l", "localFolder", true, "Defines the local folder to store the application locally. Defaults to current folder.\n"
-				+ "For each application a new folder is created automatically.");
-		option.setRequired(false);
-		option.setArgName("my/apps");
+		option.setArgName("Partners");
 		options.addOption(option);
 	}
 
 	@Override
 	public void printUsage(String message, String[] args) {
-		super.printUsage(message, args);		
-		System.out.println("You may run one of the following examples:");
-		System.out.println("Using parameters provided in properties file stored in conf-folder:");
-		System.out.println(getBinaryName()+" app get -n \"Client App\" -s api-env -o json");
-		System.out.println(getBinaryName()+" app get -n \"Client App\" -l /tmp/exported_apps -deleteTarget -s api-env -o json");
-		System.out.println(getBinaryName()+" app get -n \"App 123\" -l /tmp/exported_apps -s api-env -o json -deleteTarget");
+		super.printUsage(message, args);
+		System.out.println("----------------------------------------------------------------------------------------");
+		System.out.println("How to get/export applications with different output formats");
+		System.out.println("Get all applications on console using environment properties: env.api-env.properties:");
+		System.out.println(getBinaryName()+" app get -s api-env");
+		System.out.println("Same as before, but with output format JSON - As it is used to import applications");
+		System.out.println(getBinaryName()+" api get -s api-env -o json");
 		System.out.println();
 		System.out.println();
-		System.out.println(getBinaryName()+" app get -n \"Client App\" -h localhost -u apiadmin -p changeme -o json");
-		System.out.println(getBinaryName()+" app get -n \"App 123\" -h localhost -u apiadmin -p changeme -s prod -o json -deleteTarget");
+		System.out.println("How to filter the list of selected applications:");
+		System.out.println(getBinaryName()+" api get -s api-env -n \"Client App\" -o json");
+		System.out.println(getBinaryName()+" app get -s api-env -n \"Client App\" -t /tmp/exported_apps -o json -deleteTarget ");
+		System.out.println(getBinaryName()+" app get -s api-env -n \"App 123\" -t /tmp/exported_apps -o json -deleteTarget");
 		System.out.println();
 		System.out.println("For more information and advanced examples please visit:");
 		System.out.println("https://github.com/Axway-API-Management-Plus/apim-cli/wiki");