+++ CHANGELOG.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Feature to use a YAML based API-Definition for Swagger & OpenAPI (API-Manager 7.7 only)
 - API Backend-Server information in wide and ultra view console get view
 - Option to filter for specific backendBasepath
 - Show API detail information if only 1 API is found (See issue [#11](https://github.com/Axway-API-Management-Plus/apim-cli/issues/11))
-- New feature to load the API-Definition from FE-API (See issue [#4](https://github.com/Axway-API-Management-Plus/apim-cli/issues/4))
-- New feature to export APIs as a CSV-File
+- Feature to load the API-Definition from FE-API (See issue [#4](https://github.com/Axway-API-Management-Plus/apim-cli/issues/4))
+- Feature to export APIs as a CSV-File
+- CLI usage information for all operations improved
 
 ### Changed
 - API Console ultra view not longer renders API-Tags in detail. Only indicated with True & False